requirements.txt
# Core dependencies
httpx>=0.25.0
aiohttp>=3.9.0
asyncio>=3.4.3

# LLM Providers
openai>=1.0.0
anthropic>=0.18.0

# Web scraping and content extraction
beautifulsoup4>=4.12.0
lxml>=5.0.0
trafilatura>=1.6.0
readability-lxml>=0.8.1

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
tenacity>=8.2.0

# Development dependencies
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Type checking
mypy>=1.8.0

# Code formatting
black>=24.1.0
isort>=5.13.0
ruff>=0.2.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.5.0
mkdocstrings>=0.24.0
mkdocstrings-python>=1.8.0
